Second commit with changes made

+++ Calculator/src/main/java/operations/MultiplyOperation.java
@@ -1,15 +1,15 @@
 package operations;
 
 final public class MultiplyOperation {
-	int a = 0;
-	int b = 0;
-	public void setA(int input){
+	float a = 0;
+	float b = 0;
+	public void setA(float input){
 		this.a = input;
 	}
-	public void setB(int input){
+	public void setB(float input){
 		this.b = input;
 	}
-	public int getResult() {
+	public double getResult() {
 		return this.a * this.b;
 	}
 }
